refactor(backend): move SessionLeaveResult/SessionDisconnectResult to types.ts (review)
Per follow-up review on #3360: break the type-only circular import between
types.ts and client-lifecycle.ts by moving SessionLeaveResult and
SessionDisconnectResult into types.ts alongside RoomManagerDeps /
JoinSessionCallbacks / JoinSessionOptions, instead of importing
SessionLeaveResult back from client-lifecycle.ts. room-manager.ts now
imports both result types from ./types.
